+++ src/utils/constants/order-status-enum.ts
@@ -1,11 +1,3 @@
-export enum PaymentMethodLabel {
-  카카오페이 = '카카오페이',
-  신용카드 = '신용카드',
-  네이버페이 = '네이버페이',
-  페이코 = '페이코',
-  계좌이체 = '계좌이체',
-}
-
 export enum OrderStatusLabel {
   결제완료 = '결제완료',
   상품준비중 = '상품준비중',
