+++ src/utils/constants/payment-method-enum.ts
@@ -1,0 +1,7 @@
+export enum PaymentMethodLabel {
+  카카오페이 = '카카오페이',
+  신용카드 = '신용카드',
+  네이버페이 = '네이버페이',
+  페이코 = '페이코',
+  계좌이체 = '계좌이체',
+}
